refactor : useTotalDiscount 커스텀 훅 내의 함수 유틸로 분리
